Remove stale dirs from tests/ and move all remaining tests to the new structure

add files
