+++ tests/conftest.py
@@ -11,7 +11,7 @@ import json
 from concurrent.futures.thread import ThreadPoolExecutor
 from datetime import datetime
 from math import floor
-from shutil import copyfile
+from shutil import copyfile, rmtree
 from functools import partial
 
 import boto3
@@ -152,6 +152,7 @@ from ocs_ci.utility.utils import (
     skipif_ui_not_support,
     run_cmd,
     ceph_health_check_multi_storagecluster_external,
+    clone_repo,
 )
 from ocs_ci.helpers import helpers, dr_helpers
 from ocs_ci.helpers.helpers import (
@@ -8058,3 +8059,52 @@ def destroy_hosted_cluster():
         return destroy_res
 
     return factory
+
+
+@pytest.fixture(scope="session")
+def clone_upstream_ceph(request, tmp_path_factory):
+    """
+    fixture to make temporary directory for the 'upstream ceph' and clone repo to it
+    """
+    repo_dir = tmp_path_factory.mktemp("upstream_ceph_dir")
+
+    def finalizer():
+        rmtree(repo_dir, ignore_errors=True)
+
+    request.addfinalizer(finalizer)
+    clone_repo(
+        constants.CEPH_UPSTREAM_REPO, str(repo_dir), branch="main", tmp_repo=True
+    )
+    return repo_dir
+
+
+@pytest.fixture(scope="session")
+def clone_ocs_operator(request, tmp_path_factory):
+    """
+    fixture to make temporary directory for the 'ocs operator' and clone repo to it
+    """
+    repo_dir = tmp_path_factory.mktemp("ocs_operator_dir")
+
+    def finalizer():
+        rmtree(repo_dir, ignore_errors=True)
+
+    request.addfinalizer(finalizer)
+    clone_repo(constants.OCS_OPERATOR_REPO, str(repo_dir), branch="main", tmp_repo=True)
+    return repo_dir
+
+
+@pytest.fixture(scope="session")
+def clone_odf_monitoring_compare_tool(request, tmp_path_factory):
+    """
+    fixture to make temporary directory for the 'ODF monitor compare tool' and clone repo to it
+    """
+    repo_dir = tmp_path_factory.mktemp("monitor_tool_dir")
+
+    def finalizer():
+        rmtree(repo_dir, ignore_errors=True)
+
+    request.addfinalizer(finalizer)
+    clone_repo(
+        constants.ODF_MONITORING_TOOL_REPO, str(repo_dir), branch="main", tmp_repo=True
+    )
+    return repo_dir
