--- tests/disaster-recovery/__init__.py
@@ -1,1 +1,0 @@
-# This package is for DR team tests
