--- tests/e2e/__init__.py
@@ -1,1 +1,0 @@
-# This package is for E2E team tests
