+++ tests/functional/lvmo/test_lvm_manual_diskpath.py
