+++ tests/functional/lvmo/test_lvm_multi_snapshot.py
@@ -11,6 +11,7 @@ from ocs_ci.framework.pytest_customization.marks import (
 )
 from ocs_ci.framework.testlib import skipif_ocs_version, ManageTest, acceptance
 from ocs_ci.ocs import constants
+from ocs_ci.utility.utils import get_ocp_version
 from ocs_ci.ocs.cluster import LVM
 from ocs_ci.ocs.resources.pod import cal_md5sum
 from ocs_ci.ocs.exceptions import Md5CheckFailed
@@ -46,6 +47,7 @@ class TestLvmMultiSnapshot(ManageTest):
 
     """
 
+    ocp_version = get_ocp_version()
     pvc_size = 100
     access_mode = constants.ACCESS_MODE_RWO
     pvc_num = 5
