+++ tests/libtest/test_update_catalogsource.py
