--- tests/manage/__init__.py
@@ -1,1 +1,0 @@
-# This package is for manage team tests
